+++ include/glyphstr.h
