+++ include/meson.build
@@ -472,6 +472,10 @@ if build_xorg
             'opaque.h',
             'optionstr.h',
             'os.h',
+            'glyphstr.h',
+            'mipict.h',
+            'picture.h',
+            'picturestr.h',
             'present.h',
             'pixmap.h',
             'pixmapstr.h',
