+++ include/picturestr.h
