+++ render/meson.build
@@ -12,19 +12,8 @@ srcs_render = [
     'render.c',
 ]
 
-hdrs_render = [
-    'glyphstr.h',
-    'mipict.h',
-    'picture.h',
-    'picturestr.h',
-]
-
 libxserver_render = static_library('xserver_render',
     srcs_render,
     include_directories: inc,
     dependencies: common_dep,
 )
-
-if build_xorg
-    install_data(hdrs_render, install_dir: xorgsdkdir)
-endif
